medica/medica_bot/requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.6
cryptography==39.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.1
django-bootstrap-icons==0.8.2
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
install==1.3.5
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.4
numpy==1.23.4
oauthlib==3.2.2
openai==0.27.2
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.3.0
yarl==1.8.2
